+++ src/finlang.test/ScopeTest.cs
@@ -4,7 +4,7 @@ public class Sample
 {
     public Sample()
     {
-        math.CurrentMode.Should().Be(math.Mode.NotSpecified);
+        math.CurrentMode.Should().Be(math.Mode.Unsafe);
     }
 
     public int get_int2()
@@ -16,12 +16,12 @@ public class Sample
     public void method()
     {
         int x = get_int();
-        math.CurrentMode.Should().Be(math.Mode.NotSpecified);
+        math.CurrentMode.Should().Be(math.Mode.Unsafe);
     }
 
     public int get_int()
     {
-        math.CurrentMode.Should().Be(math.Mode.NotSpecified);
+        math.CurrentMode.Should().Be(math.Mode.Unsafe);
         return 2234;
     }
 }
@@ -31,14 +31,14 @@ public class ScopeTest
     [Fact]
     public void Test1()
     {
-        math.unsafe_mode(); // Math.scoped_pragma
-        math.CurrentMode.Should().Be(math.Mode.Unsafe);
+        math.capture_errors(new err.Err()); // Math.scoped_pragma
+        math.CurrentMode.Should().Be(math.Mode.UserProvidedErr);
 
         Sample sample = new();
-        math.CurrentMode.Should().Be(math.Mode.Unsafe);
+        math.CurrentMode.Should().Be(math.Mode.UserProvidedErr);
 
         sample.method();
-        math.CurrentMode.Should().Be(math.Mode.Unsafe);
+        math.CurrentMode.Should().Be(math.Mode.UserProvidedErr);
     }
 
     [Fact]